Added fixes for recyclerView -> Separated adapter and holder into different files.
